If neither of `value` or `encrypted_value` is present, error should be thrown Moved the EnvironmentVariableRepresenter to api-base module Removed it from api-shared-v2\v4\v6 and api-pipelineconfig-v7
