+++ api/api-base/src/test/groovy/com/thoughtworks/go/api/representers/EnvironmentVariableRepresenterTest.groovy
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv2.shared.representers
+package com.thoughtworks.go.api.representers
 
 import com.thoughtworks.go.api.util.GsonTransformer
 import com.thoughtworks.go.config.EnvironmentVariableConfig
 import com.thoughtworks.go.security.GoCipher
 import net.javacrumbs.jsonunit.fluent.JsonFluentAssert
+import org.hamcrest.MatcherAssert
 import org.junit.jupiter.api.Test
 import spark.HaltException
 
@@ -123,13 +124,13 @@ class EnvironmentVariableRepresenterTest {
     def plainVariable = environmentVariableConfig.get(1)
 
     assertThat(secureVariable.secure, is(true))
-    assertThat(secureVariable.name, is("secured"))
-    assertThat(secureVariable.value, is("confidential"))
+    MatcherAssert.assertThat(secureVariable.name, is("secured"))
+    MatcherAssert.assertThat(secureVariable.value, is("confidential"))
     assertThat(secureVariable.errors().isEmpty(), is(true))
 
     assertThat(plainVariable.secure, is(false))
-    assertThat(plainVariable.name, is("plain"))
-    assertThat(plainVariable.value, is("plaint text value"))
+    MatcherAssert.assertThat(plainVariable.name, is("plain"))
+    MatcherAssert.assertThat(plainVariable.value, is("plaint text value"))
     assertThat(plainVariable.errors().isEmpty(), is(true))
   }
