+++ api/api-environments-v2/build.gradle
@@ -19,7 +19,6 @@ apply plugin: 'groovy'
 
 dependencies {
   compile project(':api:api-base')
-  compile project(':api:api-shared-v2')
   testCompile project(path: ':api:api-base', configuration: 'testOutput')
 
   testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: project.versions.junit5
