+++ api/api-environments-v2/src/main/java/com/thoughtworks/go/apiv2/environments/EnvironmentsControllerV2.java
@@ -20,6 +20,7 @@ import com.thoughtworks.go.api.ApiController;
 import com.thoughtworks.go.api.ApiVersion;
 import com.thoughtworks.go.api.CrudController;
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper;
 import com.thoughtworks.go.api.util.GsonTransformer;
@@ -28,7 +29,6 @@ import com.thoughtworks.go.apiv2.environments.model.PatchEnvironmentRequest;
 import com.thoughtworks.go.apiv2.environments.representers.EnvironmentRepresenter;
 import com.thoughtworks.go.apiv2.environments.representers.EnvironmentsRepresenter;
 import com.thoughtworks.go.apiv2.environments.representers.PatchEnvironmentRequestRepresenter;
-import com.thoughtworks.go.apiv2.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.config.BasicEnvironmentConfig;
 import com.thoughtworks.go.config.EnvironmentConfig;
 import com.thoughtworks.go.config.EnvironmentVariableConfig;
