+++ api/api-pipeline-config-v6/src/main/java/com/thoughtworks/go/apiv6/admin/pipelineconfig/representers/PipelineConfigRepresenter.java
@@ -16,11 +16,11 @@
 package com.thoughtworks.go.apiv6.admin.pipelineconfig.representers;
 
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.ErrorGetter;
 import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.apiv6.admin.pipelineconfig.representers.materials.MaterialRepresenter;
 import com.thoughtworks.go.apiv6.admin.pipelineconfig.representers.trackingtool.TrackingToolRepresenter;
-import com.thoughtworks.go.apiv6.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.apiv6.shared.representers.configorigin.ConfigRepoOriginRepresenter;
 import com.thoughtworks.go.apiv6.shared.representers.configorigin.ConfigXmlOriginRepresenter;
 import com.thoughtworks.go.apiv6.shared.representers.stages.ConfigHelperOptions;
