+++ api/api-pipeline-config-v6/src/test/groovy/com/thoughtworks/go/apiv6/admin/pipelineconfig/representers/PipelineConfigRepresenterTest.groovy
@@ -18,7 +18,7 @@ package com.thoughtworks.go.apiv6.admin.pipelineconfig.representers
 import com.thoughtworks.go.api.util.GsonTransformer
 import com.thoughtworks.go.apiv6.admin.pipelineconfig.representers.materials.MaterialRepresenter
 import com.thoughtworks.go.apiv6.admin.pipelineconfig.representers.trackingtool.TrackingToolRepresenter
-import com.thoughtworks.go.apiv6.shared.representers.EnvironmentVariableRepresenter
+
 import com.thoughtworks.go.apiv6.shared.representers.stages.ConfigHelperOptions
 import com.thoughtworks.go.apiv6.shared.representers.stages.StageRepresenter
 import com.thoughtworks.go.config.*
