+++ api/api-pipeline-operations-v1/src/main/java/com/thoughtworks/go/apiv1/pipelineoperations/representers/PipelineScheduleOptionsRepresenter.java
@@ -15,8 +15,8 @@
  */
 package com.thoughtworks.go.apiv1.pipelineoperations.representers;
 
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.JsonReader;
-import com.thoughtworks.go.apiv6.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.server.domain.MaterialForScheduling;
 import com.thoughtworks.go.server.domain.PipelineScheduleOptions;
