+++ api/api-shared-v4/src/main/java/com/thoughtworks/go/apiv4/shared/representers/stages/JobRepresenter.java
@@ -17,10 +17,10 @@ package com.thoughtworks.go.apiv4.shared.representers.stages;
 
 import com.thoughtworks.go.api.base.OutputListWriter;
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.ErrorGetter;
 import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.apiv4.shared.representers.stages.tasks.TaskRepresenter;
-import com.thoughtworks.go.apiv4.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.apiv4.shared.representers.stages.artifacts.ArtifactRepresenter;
 import com.thoughtworks.go.config.*;
