+++ api/api-shared-v4/src/main/java/com/thoughtworks/go/apiv4/shared/representers/stages/StageRepresenter.java
@@ -17,9 +17,9 @@ package com.thoughtworks.go.apiv4.shared.representers.stages;
 
 import com.thoughtworks.go.api.base.OutputListWriter;
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.ErrorGetter;
 import com.thoughtworks.go.api.representers.JsonReader;
-import com.thoughtworks.go.apiv4.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.config.*;
 
 import java.util.HashMap;
