+++ api/api-shared-v6/src/main/java/com/thoughtworks/go/apiv6/shared/representers/stages/JobRepresenter.java
@@ -17,9 +17,9 @@ package com.thoughtworks.go.apiv6.shared.representers.stages;
 
 import com.thoughtworks.go.api.base.OutputListWriter;
 import com.thoughtworks.go.api.base.OutputWriter;
+import com.thoughtworks.go.api.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.api.representers.ErrorGetter;
 import com.thoughtworks.go.api.representers.JsonReader;
-import com.thoughtworks.go.apiv6.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.apiv6.shared.representers.stages.artifacts.ArtifactRepresenter;
 import com.thoughtworks.go.apiv6.shared.representers.stages.tasks.TaskRepresenter;
 import com.thoughtworks.go.config.*;
