+++ settings.gradle
@@ -66,7 +66,6 @@ include ':api:api-server-health-messages-v1'
 include ':api:api-server-health-v1'
 include ':api:api-server-maintenance-mode-v1'
 include ':api:api-shared-v1'
-include ':api:api-shared-v2'
 include ':api:api-shared-v4'
 include ':api:api-shared-v6'
 include ':api:api-stage-operations-v1'
